Moved WSD samples into the SAMPLES module.
