+++ samples/src/main/java/org/nanohttpd/samples/tempfiles/TempFilesServer.java
@@ -1,4 +1,4 @@
-package fi.iki.elonen;
+package org.nanohttpd.samples.tempfiles;
 
 /*
  * #%L
@@ -40,10 +40,9 @@ import java.util.List;
 import org.nanohttpd.protocols.http.tempfiles.DefaultTempFile;
 import org.nanohttpd.protocols.http.tempfiles.ITempFile;
 import org.nanohttpd.protocols.http.tempfiles.ITempFileManager;
+import org.nanohttpd.samples.http.DebugServer;
 import org.nanohttpd.util.IFactory;
 import org.nanohttpd.util.ServerRunner;
-
-import fi.iki.elonen.debug.DebugServer;
 
 /**
  * @author Paul S. Hawke (paul.hawke@gmail.com) On: 3/9/13 at 12:47 AM
