+++ samples/src/main/java/org/nanohttpd/samples/websockets/EchoSocketSample.java
@@ -1,8 +1,8 @@
-package fi.iki.elonen;
+package org.nanohttpd.samples.websockets;
 
 /*
  * #%L
- * NanoHttpd-Samples
+ * NanoHttpd-Websocket
  * %%
  * Copyright (C) 2012 - 2015 nanohttpd
  * %%
@@ -33,49 +33,23 @@ package fi.iki.elonen;
  * #L%
  */
 
-import java.util.Map;
-import java.util.logging.Logger;
+import java.io.IOException;
 
-import org.nanohttpd.protocols.http.IHTTPSession;
-import org.nanohttpd.protocols.http.NanoHTTPD;
-import org.nanohttpd.protocols.http.request.Method;
-import org.nanohttpd.protocols.http.response.Response;
-import org.nanohttpd.util.ServerRunner;
+import fi.iki.elonen.NanoWSD;
 
-/**
- * An example of subclassing NanoHTTPD to make a custom HTTP server.
- */
-public class HelloServer extends NanoHTTPD {
+public class EchoSocketSample {
 
-    /**
-     * logger to log to.
-     */
-    private static final Logger LOG = Logger.getLogger(HelloServer.class.getName());
-
-    public static void main(String[] args) {
-        ServerRunner.run(HelloServer.class);
-    }
-
-    public HelloServer() {
-        super(8080);
-    }
-
-    @Override
-    public Response serve(IHTTPSession session) {
-        Method method = session.getMethod();
-        String uri = session.getUri();
-        HelloServer.LOG.info(method + " '" + uri + "' ");
-
-        String msg = "<html><body><h1>Hello server</h1>\n";
-        Map<String, String> parms = session.getParms();
-        if (parms.get("username") == null) {
-            msg += "<form action='?' method='get'>\n" + "  <p>Your name: <input type='text' name='username'></p>\n" + "</form>\n";
-        } else {
-            msg += "<p>Hello, " + parms.get("username") + "!</p>";
+    public static void main(String[] args) throws IOException {
+        final boolean debugMode = args.length >= 2 && "-d".equals(args[1].toLowerCase());
+        NanoWSD ws = new DebugWebSocketServer(args.length > 0 ? Integer.parseInt(args[0]) : 9090, debugMode);
+        ws.start();
+        System.out.println("Server started, hit Enter to stop.\n");
+        try {
+            System.in.read();
+        } catch (IOException ignored) {
         }
-
-        msg += "</body></html>\n";
-
-        return Response.newFixedLengthResponse(msg);
+        ws.stop();
+        System.out.println("Server stopped.\n");
     }
+
 }
